requirements.txt
asgiref==3.7.2
attrs==23.2.0
backports.zoneinfo==0.2.1
Brotli==1.1.0
click==8.1.7
dj-database-url==2.1.0
Django==3.2.23
django-cookies-samesite==0.9.0
django-cors-headers==4.1.0
django-sslserver==0.22
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.2
gunicorn==21.2.0
h11==0.14.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
mysqlclient==2.1.1
packaging==24.0
pkgutil_resolve_name==1.3.10
psycopg2==2.9.9
PyJWT==2.8.0
pyrsistent==0.19.3
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
ua-parser==0.10.0
uritemplate==4.1.1
uvicorn==0.22.0
whitenoise==6.5.0
zipp==3.15.0
